Add loading screens and refactor Merge component
- Introduced loading screens to provide real-time feedback during the merge operation.
- Refactored the Merge component for better code clarity and separation of concerns.
- Implemented rounding off of progress values to two decimal places.
- Optimized state management related to progress updates.
