+++ backend/pdf-transformation-monorepo/pdf-merger-service/src/main/java/com/jakubrurak/apigateway/ApiGatewayResponse.java
@@ -1,6 +1,5 @@
 package com.jakubrurak.apigateway;
 
-import java.util.List;
 import java.util.Map;
 
 public class ApiGatewayResponse {
