+++ frontend/src/components/PdfTransformationApp.css
@@ -1,0 +1,23 @@
+html, body, main {
+    height: 100%;
+    margin: 0;
+    padding: 0;
+    font-family: 'Arial', sans-serif;
+}
+
+body {
+    display: flex;
+    flex-direction: column;
+}
+
+main {
+    flex: 1 0 auto; /* It grows to fill the space but doesn't shrink. */
+}
+
+main {
+    min-height: 100vh;
+}
+
+.FooterComponent { 
+    flex-shrink: 0; 
+}
